splitup vector_layer
