+++ src/actinia_core/rest/vector_layer.py
@@ -4,7 +4,7 @@
 # performance processing of geographical data that uses GRASS GIS for
 # computational tasks. For details, see https://actinia.mundialis.de/
 #
-# Copyright (c) 2016-2018 Sören Gebbert and mundialis GmbH & Co. KG
+# Copyright (c) 2016-2022 Sören Gebbert and mundialis GmbH & Co. KG
 #
 # This program is free software: you can redistribute it and/or modify
 # it under the terms of the GNU General Public License as published by
@@ -28,29 +28,24 @@ from flask import jsonify, make_response, request
 from flask_restful_swagger_2 import swagger
 import os
 import pickle
-from shutil import rmtree
 from uuid import uuid4
 from werkzeug.utils import secure_filename
 from zipfile import ZipFile
 from actinia_api.swagger2.actinia_core.schemas.vector_layer import \
-    VectorInfoResponseModel, VectorRegionCreationModel, \
-    VectorAttributeModel, VectorInfoModel
+    VectorInfoResponseModel, VectorRegionCreationModel
 
 from actinia_core.core.common.app import URL_PREFIX
 from actinia_core.core.common.redis_interface import enqueue_job
-from actinia_core.core.common.exceptions import AsyncProcessError
 from actinia_core.core.utils import allowed_file
 from actinia_core.models.response_models import \
     ProcessingResponseModel, ProcessingErrorResponseModel, SimpleResponseModel
-from actinia_core.processing.actinia_processing.ephemeral.ephemeral_processing \
-     import EphemeralProcessing
-from actinia_core.processing.actinia_processing.ephemeral.persistent_processing \
-     import PersistentProcessing
 from actinia_core.rest.base.map_layer_base import MapLayerRegionResourceBase
+from actinia_core.processing.common.vector_layer import \
+    start_create_job, start_delete_job, start_info_job
 
 __license__ = "GPLv3"
 __author__ = "Sören Gebbert, Carmen Tawalika, Guido Riembauer, Anika Weinmann"
-__copyright__ = "Copyright 2016-2021, Sören Gebbert and mundialis GmbH & Co. KG"
+__copyright__ = "Copyright 2016-2022, Sören Gebbert and mundialis GmbH & Co. KG"
 __maintainer__ = "mundialis"
 
 
@@ -310,197 +305,3 @@ class VectorLayerResource(MapLayerRegionResourceBase):
 
         http_code, response_model = pickle.loads(self.response_data)
         return make_response(jsonify(response_model), http_code)
-
-
-def start_info_job(*args):
-    processing = EphemeralVectorInfo(*args)
-    processing.run()
-
-
-class EphemeralVectorInfo(EphemeralProcessing):
-
-    def __init__(self, *args):
-        EphemeralProcessing.__init__(self, *args)
-        self.response_model_class = VectorInfoResponseModel
-
-    def _execute(self):
-        """Read info from a vector layer
-
-        Use a temporary mapset for processing
-        """
-        self._setup()
-
-        vector_name = self.rdc.map_name
-        self.required_mapsets.append(self.mapset_name)
-
-        pc = {}
-        pc["1"] = {
-            "module": "v.info",
-            "inputs": {"map": vector_name + "@" + self.mapset_name},
-            "flags": "gte"}
-
-        pc["2"] = {
-            "module": "v.info",
-            "inputs": {"map": vector_name + "@" + self.mapset_name},
-            "flags": "h"}
-
-        pc["3"] = {
-            "module": "v.info",
-            "inputs": {"map": vector_name + "@" + self.mapset_name},
-            "flags": "c"}
-
-        self.skip_region_check = True
-        process_list = self._create_temporary_grass_environment_and_process_list(
-            process_chain=pc, skip_permission_check=True)
-        self._execute_process_list(process_list)
-
-        kv_list = self.module_output_log[0]["stdout"].split("\n")
-
-        vector_info = {}
-        # Regular metadata
-        for string in kv_list:
-            if "=" in string:
-                k, v = string.split("=", 1)
-                vector_info[k] = v
-
-        kv_list = self.module_output_log[1]["stdout"].split("\n")
-        # Command that created the vector
-        for string in kv_list:
-            if "COMMAND:" in string:
-                k, v = string.split(":", 1)
-                vector_info[k] = v
-
-        datatypes = self.module_output_log[2]["stdout"].split("\n")
-
-        # Datatype of the vector table
-        attr_list = []
-        for string in datatypes:
-            if "|" in string:
-                dt_dict = {}
-                col_type, col_name = string.split("|", 1)
-                dt_dict["type"] = col_type
-                dt_dict["column"] = col_name
-                attr_list.append(VectorAttributeModel(**dt_dict))
-
-        vector_info["Attributes"] = attr_list
-
-        self.module_results = VectorInfoModel(**vector_info)
-
-
-def start_delete_job(*args):
-    processing = PersistentVectorDeleter(*args)
-    processing.run()
-
-
-class PersistentVectorDeleter(PersistentProcessing):
-
-    def __init__(self, *args):
-        PersistentProcessing.__init__(self, *args)
-
-    def _execute(self):
-        """Delete a specific vector layer from a location in the user database
-
-        Use the original mapset for processing
-        """
-        self._setup()
-
-        vector_name = self.map_name
-        self.required_mapsets.append(self.target_mapset_name)
-
-        pc = {}
-        pc["1"] = {"module": "g.remove", "inputs": {"type": "vector",
-                                                    "name": vector_name},
-                   "flags": "f"}
-
-        self.skip_region_check = True
-        process_list = self._validate_process_chain(process_chain=pc,
-                                                    skip_permission_check=True)
-        self._check_lock_target_mapset()
-        self._create_grass_environment(grass_data_base=self.grass_user_data_base,
-                                       mapset_name=self.target_mapset_name)
-
-        self._execute_process_list(process_list)
-
-        if "WARNING: No data base element files found" in "\n".join(
-                self.module_output_log[0]["stderr"]):
-            raise AsyncProcessError("Vector layer <%s> not found" % (vector_name))
-
-        self.finish_message = "Vector layer <%s> successfully removed." % vector_name
-
-
-def start_create_job(*args):
-    processing = PersistentVectorCreator(*args)
-    processing.run()
-
-
-class PersistentVectorCreator(PersistentProcessing):
-
-    def __init__(self, *args):
-
-        PersistentProcessing.__init__(self, *args)
-
-    def _execute(self):
-        """Create a specific vector layer
-
-        This approach is complex, since the vector generation is performed in a local
-        temporary mapset that is later merged into the target mapset. Workflow:
-
-        1. Check the process chain
-        2. Lock the temp and target mapsets
-        3. Setup GRASS and create the temporary mapset
-        4. Execute g.list of the first process chain to check if the target
-           vector exists
-        5. If the target vector does not exist then run v.import
-        6. Copy the local temporary mapset to the storage and merge it into the
-           target mapset
-        """
-        self._setup()
-
-        vector_name = self.map_name
-        self.required_mapsets.append(self.target_mapset_name)
-
-        pc_1 = {}
-        pc_1["1"] = {"module": "g.list", "inputs": {"type": "vector",
-                                                    "pattern": vector_name,
-                                                    "mapset": self.target_mapset_name}}
-        # Check the first process chain
-        self.skip_region_check = True
-        pc_1 = self._validate_process_chain(skip_permission_check=True,
-                                            process_chain=pc_1)
-
-        pc_2 = {}
-        pc_2["1"] = {"module": "v.import",
-                     "inputs": {"input": self.rdc.request_data},
-                     "outputs": {"output": {"name": vector_name}}}
-        # Check the second process chain
-        self.skip_region_check = True
-        pc_2 = self._validate_process_chain(skip_permission_check=True,
-                                            process_chain=pc_2)
-
-        self._check_lock_target_mapset()
-        self._lock_temp_mapset()
-        self._create_temporary_grass_environment(
-            source_mapset_name=self.target_mapset_name)
-        self._execute_process_list(pc_1)
-
-        # check if vector exists
-        raster_list = self.module_output_log[0]["stdout"].split("\n")
-
-        if len(raster_list[0]) > 0:
-            raise AsyncProcessError("Vector layer <%s> exists." % vector_name)
-
-        self._execute_process_list(pc_2)
-        self._copy_merge_tmp_mapset_to_target_mapset()
-
-        # Delete imported file
-        msg = ""
-        try:
-            if self.rdc.request_data.endswith('.shp'):
-                rmtree(os.path.dirname(self.rdc.request_data), ignore_errors=True)
-            else:
-                os.remove(self.rdc.request_data)
-        except Exception:
-            msg = " WARNING: Uploaded file cannot be removed."
-
-        self.finish_message = (f"Vector layer <{vector_name}> successfully "
-                               f"imported.{msg}")
